discover: Allow discover server to receive boot messages
Currently, the petitboot socket is one way: messages are only sent from
server to client.

Beause we want the clients to trigger a boot, this change allows the
server to receive messages on the petitboot socket. This is just a
matter of adding a waiter to the client-specific socket, then handling
incoming messages.

At present we don't do anything with the messages, but we'll add a
handler later.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>