test: Add URL parser test infrastructure
This change adds some simple testing infrastrcture to the URL parser.
We use a small C binary (parse-url) to run the url parser on its
argument, and compare the output with an expected datafile.

An initial test is included, to check the behaviour of URLs with
multiple slashes between host and pathname. This test currently fails.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Geoff Levand <geoff@infradead.org>