discover/yaboot: fix device parsing
A couple of fixes for yaboot's device-handling code. Firstly, we need to
use 'device=' rather than 'root=', as the latter is purely for ybin, to
define where the yaboot binary goes.

Secondly, we need to respect global and option-specific device=
parameters. To do this, we keep all boot_image and initrd strings in the
state, and create the actual resources in yaboot_finish.

Add a test for all override cases, and fix the incorrect boot= parsing
in the rh8 test.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>