ui/common: Don't talloc_free sysinfo and status explicitly
Currently, we talloc_free() the sysinfo and status messages in the
discover client, right after performing the ui's callback.

However, the client may wish to keep a reference to this information
beyond the callback invocation. This change introduces a per-message
talloc context, which these structures are allocated under. This allows
the callbacks to talloc_steal() anything they wish to keep a reference
to.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>