pb-config/powerpc: fix default option handling
Currently, we don't update nvram if an NVRAM parameter is set to the
default. This means we can never revert a configuration to its default
value.

This change fixes the default setting behaviour; instead of checking for
a default, we want to check if it's a default and the option is absent.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>