lib/pb-config: Add config_copy
At the moment, UIs have the config_set_defaults function to estabilish
an initial configuration when performing an update.

Rather than using the defaults, this change provides a config_copy()
function, so that the updated configuration can be initialised from the
current config.

With this in place, the UI/server-common pb-config module can be reduced
to just the one function.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>