discover: Call mount syscall directly
We used to use the mount binary to do filesystem autodetection. Since we
now know the fstype, we may as well call the mount syscall directly.

We add a log messages too, as we'll no longer get the 'running process:'
output from the process code, which is helpful is debugging discovery
issues.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>