discover: Fix uninitialised var warnings
We get a couple of uninitialised var warning when compiling with certain
CFLAGS (-fprofile-arcs -ftest-coverage at this stage).

In statement_if_execute: We'll never actually use this uninitialised (as
there must be at least one conditional in the parsed statement), but we
should address the warning nonetheless.

As passed to strtok_r: strtok will initialise this, but it isn't obvious
to the compiler.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>