discover: allow separate lifetimes for network interfaces and discover devices
We want the network code's network interfaces to (potentially) persist
remove events. For example, discover devices may be removed by a user
event (this happens during a udhcpc deconfig). In this case, we want the
boot options to be removed, but the struct interface needs to stay
present.

This change adds network_(un)_register_device functions, to allow the
device handler to detach from and attach to interfaces.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>