ui/ncurses: Fix clearing a boot device UUID
Currently, we have a bug where we can't clear a boot device UUID, as we
only update config->boot_device if we're in AUTOBOOT_ONE state. This
means that the config UI can't be used to change out of "autoboot from a
single device" mode.

This sets the autoboot device unconditionally, allowing it to be
cleared.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>