discover/pxe: Format IPAPPEND mac addresses correctly
The SYSAPPEND/IPAPPEND option 2 in PXE configs requires
the MAC address of the booting interface to be appended
to the boot options. Previously we formatted this as

	"BOOTIF=01:02:03:04:05:06",

but syslinux/pxelinux implementation use this format:

	"BOOTIF=01-01-02-03-04-05-06",

where the leading '01' represents the hardware type.

The relevant part of the pxelinux doc is at:

http://www.syslinux.org/wiki/index.php/SYSLINUX#SYSAPPEND_bitmask

Signed-off-by: Samuel Mendoza-Jonas <sam.mj@au1.ibm.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>