pb-plugin: Update to chroot-style plugins
This change uses a chroot for all plugins, so that plugins have complete
flexibility with their libraries, dependencies and configuration.

We remove the 'install' action, as we simply run the plugin once.
Running involves extracting the archive, setting up a root filesystem,
and running a chroot.

To simplify plugin discovery behaviour, we standardise the plugin file
to be at pb-plugin.conf of attached devices, and read the metadatafile
straight out of the archive.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>