pb-plugin: Only require a fixed plugin extension, instead of full name
Currently, the `pb-plugin scan` code requires a fixed archive name at
the root of distributed media. This means that vendors need to have a
fixed name, and can only ship one plugin per device.

This change shifts the requirement to just a filename extension
(.pb-plugin) instead.

To handle this, we now look for any plugin with a .pb-plugin extension,
and so may find multiple plugins on a device.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>