pb-plugin: Implement plugin ABI versioning
If a future update to plugins is not compatbile with the installed
plugin infrastructure, we want to abort the scan / install process.

This change adds an ABI version to the script, as well as version (and
minimum-supported version) to the plugin metadata. This means we can
check before plugin execution.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>