discover/ipmi: Use advisory locking on ipmi device
The ipmi module is shared between pb-config and pb-discover. During
system initilisation, we're likely to have a few invocations of
pb-config running, as well as pb-discover starting.

We may have multiple potential concurrent accesses to the ipmi device
node. To avoid problems, this change introduces a fcntl lock on the
device node.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>