discover: Add support for IPMI network override
On BMC platforms the 'Get System Boot Options' command can also be used
to check for a temporary network interface config override. This is
implemented via the optional 'OEM Parameters' field defined in the IPMI
v2 spec. We define the actual format of the field as:
	- 4 byte cookie value
	- 2 byte version value
	- 1 byte hardware address size
	- 1 byte IP address size
	- Hardware address
	- 1 byte flags for 'ignore' and 'method'
	And for static configs:
	- IP Address
	- 1 byte subnet value
	- Gateway address

If set the config override replaces any other interface config, forcing
the use of the specified configuration.

Signed-off-by: Sam Mendoza-Jonas <sam@mendozajonas.com>